# docs/swagger.yaml
basePath: /ih/mobile_central_2_0_0
definitions:
...
  dto.AppDownloadedUserDto:
    properties:
      machineId:
        type: string
      playerId:
        type: string
    required:
    - machineId
    - playerId
    type: object
...
  dto.DeleteAccountDto:
    properties:
      contactNumber:
        type: string
      dob:
        type: string
      documentNumber:
        type: string
      email:
        type: string
      fullname:
        type: string
      prn:
        type: string
    required:
    - dob
    - documentNumber
    - fullname
    - prn
    type: object
...
  dto.FeedbackDto:
    properties:
      accountNo:
        type: string
      apptSchedulingRating:
        type: integer
      feedbackDesc:
        type: string
      feedbackFiles:
        items:
          $ref: '#/definitions/dto.FeedbackFileDto'
        type: array
      foodBeveragesRating:
        type: integer
      hospitalServiceRating:
        type: integer
      overallRating:
        type: integer
      patientPrn:
        type: string
      paymentBillingRating:
        type: integer
      recommendUsRating:
        type: integer
      staffServiceRating:
        type: integer
      visitType:
        type: string
    required:
    - accountNo
    - apptSchedulingRating
    - feedbackDesc
    - foodBeveragesRating
    - hospitalServiceRating
    - overallRating
    - patientPrn
    - paymentBillingRating
    - recommendUsRating
    - staffServiceRating
    - visitType
    type: object
  dto.FeedbackFileDto:
    properties:
      filename:
        type: string
      mimeType:
        type: string
    required:
    - filename
    - mimeType
    type: object
...
  feedback.Feedback:
    properties:
      accountNo:
        type: string
      apptSchedulingRating:
        type: integer
      feedback_id:
        type: integer
      feedbackDesc:
        type: string
      foodBeveragesRating:
        type: integer
      hasAttachment:
        type: boolean
      hospitalServiceRating:
        type: integer
      overallRating:
        type: integer
      patientPrn:
        type: string
      paymentBillingRating:
        type: integer
      recommendUsRating:
        type: integer
      staffServiceRating:
        type: integer
      submittedDateTime:
        type: string
      submittedDateTimeExcel:
        type: string
      visitType:
        type: string
    type: object
  feedback.FeedbackAttachment:
    properties:
      attachmentContent:
        type: string
      attachmentFilename:
        type: string
      attachmentType:
        type: string
      feedback_attachment_id:
        type: integer
      feedback_id:
        type: integer
    type: object
  futureOrder.FutureOrder:
    properties:
      description:
        type: string
      orderDoctor:
        type: string
      patientName:
        type: string
      performDate:
        type: string
      planType:
        type: string
      prn:
        type: string
    type: object
...
info:
  contact: {}
  description: Vesalius-m Backend API Docs (IHP-UAT Version).
  title: (IHP-UAT) Vesalius-m Backend API
  version: "1.0"
paths:
...
  /common/downloaded-app-v2:
    post:
      parameters:
      - description: AppDownloadedUserDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AppDownloadedUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - Common
  /common/downloaded-app/{playerId}:
    post:
      parameters:
      - description: playerId
        in: path
        name: playerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - Common
...
  /common/request-delete-account:
    post:
      consumes:
      - application/json
      parameters:
      - description: DeleteAccountDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteAccountDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - Common
...
  /feedback:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Files
        in: formData
        name: files
        type: file
      - description: FeedbackDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FeedbackDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - Feedback
  /feedback/{feedbackId}:
    get:
      parameters:
      - description: feedbackId
        in: path
        name: feedbackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/feedback.Feedback'
      security:
      - BearerAuth: []
      tags:
      - Feedback
  /feedback/all:
    get:
      parameters:
      - description: _page
        in: query
        name: _page
        type: string
      - description: _limit
        in: query
        name: _limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/feedback.FeedbackAttachment'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Feedback
  /feedback/attachment-download/{attachmentId}:
    get:
      parameters:
      - description: attachmentId
        in: path
        name: attachmentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      tags:
      - Feedback
  /feedback/attachment/{feedbackId}:
    get:
      parameters:
      - description: feedbackId
        in: path
        name: feedbackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/feedback.FeedbackAttachment'
      security:
      - BearerAuth: []
      tags:
      - Feedback
  /feedback/export:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/feedback.Feedback'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Feedback
  /future-order/all/{prn}:
    get:
      parameters:
      - description: prn
        in: path
        name: prn
        required: true
        type: string
      - description: _page
        in: query
        name: _page
        type: string
      - description: _limit
        in: query
        name: _limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/futureOrder.FutureOrder'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Future Order
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"